pyproject.toml
[tool.poetry]
name = "todo-list"
version = "0.1.0"
description = ""
authors = ["Bryan Ang <bryanang2001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
routes = "^2.5.1"
flask = "^3.0.0"
pymongo = "^4.6.1"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
